requirements.txt
astropy>=4.2.1
matplotlib>=3.4.2
pandas>==1.2.4
numpy>=1.21.1
torch>=1.7.1
tqdm>=4.61.0
seaborn>=0.11.1
scikit_image>=0.18.1
torchvision>=0.8.2
Pillow>=8.3.1
scikit_learn>=0.24.2